% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comment_golem_favicon}
\alias{comment_golem_favicon}
\title{Comment out favicon}
\usage{
comment_golem_favicon(path)
}
\description{
In app_ui.R, golem uses favicon,
which does not work well with webR,
likely because of the fs package (to check).
This function will comment this line so the app UI
prints correctly.
}
\keyword{internal}
