% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{edit_app_sys}
\alias{edit_app_sys}
\title{Edit golem internal app_sys}
\usage{
edit_app_sys(path)
}
\description{
app_sys calls system.file on the local
package which can't work because the pkg
can't be install locally due to webR restrictions.
We overwrite it by a simple function which returns
the provided path prefixed by inst, where app assets
are found in a golem app.
}
\keyword{internal}
