% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_pkg_imports}
\alias{find_pkg_imports}
\title{Find package imports}
\usage{
find_pkg_imports(path)
}
\description{
Read the DESCRIPTION file and look for
any imports. Returns a vector of R dependencies.
}
\keyword{internal}
