% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold-webr.R
\name{init_shiny_webr}
\alias{init_shiny_webr}
\title{Create a webR compatible scaffold}
\usage{
init_shiny_webr(path = "./webr", copy_app_files = TRUE)
}
\arguments{
\item{path}{Path to setup the webR installation.}

\item{copy_app_files}{Whether to copy the Shiny app package main
assets (R folder, inst, NAMESPACE and DESCRIPTION). Default
to TRUE. If FALSE, it will only copy the webR JS assets and html
elements.}
}
\description{
Create a webR compatible scaffold
}
