% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold-webr.R
\name{remove_shiny_webr}
\alias{remove_shiny_webr}
\title{Cleanup the webR installation}
\usage{
remove_shiny_webr(path = "./webr")
}
\arguments{
\item{path}{Path to cleanup containing the webR installation.}
}
\description{
Cleanup the webR installation
}
