% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold-webr.R
\name{reset_shiny_webr}
\alias{reset_shiny_webr}
\title{Cleanup the webR installation}
\usage{
reset_shiny_webr(path)
}
\arguments{
\item{path}{Path to cleanup containing the webR installation.}
}
\description{
Cleanup the webR installation
}
