% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold-webr.R
\name{scaffold_webr}
\alias{scaffold_webr}
\title{Create a webR compatible scaffold}
\usage{
scaffold_webr(path)
}
\arguments{
\item{path}{Shiny app path.}
}
\description{
Create a webR compatible scaffold
}
