% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_app_deps}
\alias{set_app_deps}
\title{Get app package dependencies to webr-shiny.js}
\usage{
set_app_deps(path, conn)
}
\arguments{
\item{path}{Path containing the Shiny webR installation.}

\item{conn}{webr-shiny.js content.}
}
\description{
This allows to programmatically
write all the necessaries app dependencies to
the JS file and load them with library. This is necessary
because we can't install the local package from source so
we can't benefit from the NAMESPACE (imports).
}
\keyword{internal}
