% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_app_files}
\alias{set_app_files}
\title{Get app files within the webR installation}
\usage{
set_app_files(path, conn)
}
\arguments{
\item{path}{Path containing the Shiny webR installation.}

\item{conn}{webr-shiny.js content.}
}
\description{
This allows to programmatically
generate a list of file to write in
the webR virtual file system
}
\keyword{internal}
