% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_app_files_to_js}
\alias{write_app_files_to_js}
\title{Write app files to webr-shiny.js}
\usage{
write_app_files_to_js(path)
}
\arguments{
\item{path}{Path containing the Shiny webR installation.}
}
\description{
This allows to programmatically
generate a list of file to write in
the webR virtual file system
}
\keyword{internal}
