% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_webr_js}
\alias{write_webr_js}
\title{Write shiny-webr.js file}
\usage{
write_webr_js(path)
}
\arguments{
\item{path}{Path containing the Shiny webR installation.}
}
\description{
Write app deps and files with  \link{set_app_deps} and
\link{set_app_files}.
}
\keyword{internal}
